# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Books API",
    "version": "1.0.0",
    "description": "A simple API for working with books"
  },
  "servers": [
    {
      "url": "/",
      "description": "Current server"
    }
  ],
  "paths": {
    "/books": {
      "get": {
        "summary": "Get all books",
        "tags": [
          "Books"
        ],
        "responses": {
          "200": {
            "description": "Books returned successfully"
          },
          "500": {
            "description": "Unable to retrieve books"
          }
        }
      },
      "post": {
        "summary": "Create a book",
        "tags": [
          "Books"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "id",
                  "authorId",
                  "title",
                  "publicationDate"
                ],
                "properties": {
                  "id": {
                    "type": "string"
                  },
                  "authorId": {
                    "type": "string"
                  },
                  "title": {
                    "type": "string"
                  },
                  "publicationDate": {
                    "type": "string"
                  }
                }
              },
              "example": {
                "id": "b4",
                "authorId": "a1",
                "title": "A New Book",
                "publicationDate": "2024-01-15"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Book created successfully"
          },
          "400": {
            "description": "Missing fields, duplicate ID, or invalid author ID"
          },
          "500": {
            "description": "Unable to create book"
          }
        }
      }
    },
    "/books/{id}": {
      "get": {
        "summary": "Get one book by ID",
        "tags": [
          "Books"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "description": "The custom book ID, such as b1",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Book returned successfully"
          },
          "404": {
            "description": "Book not found"
          },
          "500": {
            "description": "Unable to retrieve book"
          }
        }
      },
      "put": {
        "summary": "Update a book by ID",
        "tags": [
          "Books"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "description": "The custom book ID, such as b1",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "authorId",
                  "title",
                  "publicationDate"
                ],
                "properties": {
                  "authorId": {
                    "type": "string"
                  },
                  "title": {
                    "type": "string"
                  },
                  "publicationDate": {
                    "type": "string"
                  }
                }
              },
              "example": {
                "authorId": "a2",
                "title": "Updated Book Title",
                "publicationDate": "2024-02-20"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Book updated successfully"
          },
          "400": {
            "description": "Missing fields or invalid author ID"
          },
          "404": {
            "description": "Book not found"
          },
          "500": {
            "description": "Unable to update book"
          }
        }
      },
      "delete": {
        "summary": "Delete a book by ID",
        "tags": [
          "Books"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "description": "The custom book ID, such as b1",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Book deleted successfully"
          },
          "404": {
            "description": "Book not found"
          },
          "500": {
            "description": "Unable to delete book"
          }
        }
      }
    },
    "/authors": {
      "get": {
        "summary": "Get all authors",
        "tags": [
          "Authors"
        ],
        "responses": {
          "200": {
            "description": "A list of authors"
          },
          "500": {
            "description": "Unable to retrieve authors"
          }
        }
      },
      "post": {
        "summary": "Create an author",
        "tags": [
          "Authors"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "id",
                  "name",
                  "birthYear"
                ],
                "properties": {
                  "id": {
                    "type": "string"
                  },
                  "name": {
                    "type": "string"
                  },
                  "birthYear": {
                    "type": "number"
                  }
                }
              },
              "example": {
                "id": "a4",
                "name": "Example Author",
                "birthYear": 1980
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Author created successfully"
          },
          "400": {
            "description": "Missing required fields or author ID already exists"
          },
          "500": {
            "description": "Unable to create author"
          }
        }
      }
    },
    "/authors/{id}": {
      "get": {
        "summary": "Get author by ID",
        "tags": [
          "Authors"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The custom author ID, such as a1"
          }
        ],
        "responses": {
          "200": {
            "description": "A single author"
          },
          "404": {
            "description": "Author not found"
          },
          "500": {
            "description": "Unable to retrieve author"
          }
        }
      },
      "put": {
        "summary": "Update an author by ID",
        "tags": [
          "Authors"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The custom author ID, such as a1"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name",
                  "birthYear"
                ],
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "birthYear": {
                    "type": "number"
                  }
                }
              },
              "example": {
                "name": "Updated Author",
                "birthYear": 1981
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Author updated successfully"
          },
          "400": {
            "description": "Missing required fields"
          },
          "404": {
            "description": "Author not found"
          },
          "500": {
            "description": "Unable to update author"
          }
        }
      },
      "delete": {
        "summary": "Delete an author by ID",
        "tags": [
          "Authors"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The custom author ID, such as a1"
          }
        ],
        "responses": {
          "204": {
            "description": "Author deleted successfully"
          },
          "404": {
            "description": "Author not found"
          },
          "409": {
            "description": "Author cannot be deleted because books still reference the author"
          },
          "500": {
            "description": "Unable to delete author"
          }
        }
      }
    }
  },
  "components": {},
  "tags": []
}